spring/orm/JPA/Biblioteca/src/main/java/com/curso/java/entidades/minimo/EntidadMinima.java
/*
 * To change this template, choose Tools | Templates
 * and open the template in the editor.
 */
package com.curso.java.entidades.minimo;

import java.io.Serializable;
import javax.persistence.Entity;
import javax.persistence.Id;

/**
 *
 * @author jose maria
 */
@Entity
public class EntidadMinima implements Serializable {

    private static final long serialVersionUID = 1L;
    @Id
    private Long id;
    private String nombre;

    public String getNombre() {
        return nombre;
    }

    public void setNombre(String nombre) {
        this.nombre = nombre;
    }

    public Long getId() {
        return id;
    }

    public void setId(Long id) {
        this.id = id;
    }

    @Override
    public int hashCode() {
        int hash = 0;
        hash += (id != null ? id.hashCode() : 0);
        return hash;
    }

    @Override
    public boolean equals(Object object) {
        // TODO: Warning - this method won't work in the case the id fields are not set
        if (!(object instanceof EntidadMinima)) {
            return false;
        }
        EntidadMinima other = (EntidadMinima) object;
        return (this.id != null || other.id == null) && (this.id == null || this.id.equals(other.id));
    }

    @Override
    public String toString() {
        return "com.curso.java.entidades.EntidadMinima[ id=" + id + " ]";
    }
}
